requirements.txt
Click==7.0
Jinja2==2.10.3
pydantic==1.0
PyYAML==5.1.2
